Add DaisyUI styling. Restructure files
