+++ apps/adventure/package.json
@@ -7,23 +7,27 @@
     "dev": "vite",
     "build": "tsc && vite build",
     "preview": "vite preview",
-    "build:tailwind": "tailwindcss -i src/input.css -o src/assets/output.css"
+    "build:tailwind": "tailwindcss -i src/input.css -o public/output.css",
+    "watch:tailwind": "pnpm run build:tailwind -- --watch"
   },
   "dependencies": {
-    "@headlessui/react": "^1.7.3",
-    "@heroicons/react": "^2.0.12",
     "@lauf/stepmachine": "workspace:^0.1.0",
     "@lauf/store": "workspace:^1.1.1",
     "@lauf/store-edit": "workspace:^1.1.1",
     "@lauf/store-react": "workspace:^1.1.1",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "tailwindcss": "^3.2.1"
+    "react-daisyui": "^2.4.4",
+    "react-dom": "^18.2.0"
   },
   "devDependencies": {
+    "@tailwindcss/typography": "^0.5.7",
     "@types/react": "^18.0.17",
     "@types/react-dom": "^18.0.6",
     "@vitejs/plugin-react": "^2.1.0",
+    "autoprefixer": "^10.4.12",
+    "daisyui": "^2.33.0",
+    "postcss": "^8.4.18",
+    "tailwindcss": "^3.2.1",
     "typescript": "^4.6.4",
     "vite": "^3.1.0"
   }
